requirements.txt
SQLAlchemy
psycopg2-binary
websockets
python-dotenv
jsonschema
python-dotenv

redis
networkx
PyJWT
structlog

Flask
Flask-CORS
aiohttp

requests
pyjwt
cryptography